+++ src/shiftedarray.jl
@@ -73,8 +73,11 @@ ShiftedVector(v::AbstractVector, n = (0,); default = missing) = ShiftedArray(v, 
 
 size(s::ShiftedArray) = size(parent(s))
 
+# Computing a shifted index (subtracting the offset)
+offset(offsets::NTuple{N,Int}, inds::NTuple{N,Int}) where {N} = map(-, inds, offsets)
+
 @inline function getindex(s::ShiftedArray{<:Any, <:Any, N}, x::Vararg{Int, N}) where {N}
-    i = OffsetArrays.offset(shifts(s), x)
+    i = offset(shifts(s), x)
     v = parent(s)
     if checkbounds(Bool, v, i...)
         @inbounds ret = v[i...]