+++ test/runtests.jl
@@ -1,5 +1,4 @@
-using ShiftedArrays, OffsetArrays
-using Test
+using ShiftedArrays, Test
 
 @testset "ShiftedVector" begin
     v = [1, 3, 5, 4]
@@ -97,29 +96,4 @@ end
     @test isequal(diff2, [-7, -9, missing, missing])
 
     @test all(lead(v, 2, default = -100) .== coalesce.(lead(v, 2), -100))
-end
-
-@testset "reduce" begin
-    v = [1, 3, 5, 6, 7, 8, 9, 11]
-    ss = ShiftedArray.((v,), [-1, -3, -6])
-    @test reduce(+, ss, -1:2) == [10, 14, 18, 23]
-    @test mapreduce(t -> t^2, +, ss, -1:2) == [58, 90, 126, 195]
-    @test isequal(reduce(+, ss, -10:2),
-     [missing, missing, missing, missing, missing, 1, 3, 5, 7, 10, 14, 18, 23])
-end
-
-@testset "reduce_vec" begin
-    v = [1, 3, 5, 6, 7, 8, 9, 11]
-    ss = ShiftedArray.((v,), [-1, -3, -6])
-    @test reduce_vec(sum, ss, -1:2) == [10, 14, 18, 23]
-    @test mapreduce_vec(t -> t^2, sum, ss, -1:2) == [58, 90, 126, 195]
-    @test isequal(reduce_vec(sum, ss, -10:2),
-     [missing, missing, missing, missing, missing, 1, 3, 5, 7, 10, 14, 18, 23])
-end
-
-@testset "offset" begin
-    v = [1, 3, 5, 6, 7, 8, 9, 11]
-    ss = ShiftedArray.((v,), [-1, -3, -6])
-    @test isequal(ShiftedArrays.to_array(ss, -1:2), [missing 3 7; 1 5 8; 3 6 9; 5 7 11])
-    @test isequal(ShiftedArrays.to_offsetarray(ss, -1:2), OffsetArray([missing 3 7; 1 5 8; 3 6 9; 5 7 11], -1:2, 1:3))
 end